requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
Pillow==9.4.0
pymongo==4.3.3
pytz==2022.7.1
sqlparse==0.2.4
